+++ src/pages/todo.tsx
@@ -1,8 +1,8 @@
-import { useState } from 'react'
-import { useRecoilValue, useSetRecoilState, selector } from 'recoil'
+import { useRecoilValue, selector } from 'recoil'
 import TodoItem from '../components/todoItem'
 import TodoListStats from '../components/todoStats'
 import TodoListFilter from '../components/todoFilter'
+import TodoItemCreator from '../components/todoCreater'
 import { todoListState, todoListFilterState } from '../atoms/todo'
 
 const TodoList = () => {
@@ -19,39 +19,6 @@ const TodoList = () => {
       ))}
     </div>
   )
-}
-const TodoItemCreator = () => {
-  const [inputValue, setInputValue] = useState('')
-  const setTodoList = useSetRecoilState(todoListState)
-
-  const addItem = () => {
-    setTodoList((oldTodoList) => [
-      ...oldTodoList,
-      {
-        id: getId(),
-        text: inputValue,
-        isComplete: false,
-      },
-    ])
-    setInputValue('')
-  }
-
-  const onChange = ({ target: { value } }) => {
-    setInputValue(value)
-  }
-
-  return (
-    <div>
-      <input type="text" className="" value={inputValue} onChange={onChange} />
-      <button onClick={addItem}>Add</button>
-    </div>
-  )
-}
-
-// utility for creating unique Id
-let id = 0
-function getId() {
-  return id++
 }
 
 const filteredTodoListState = selector({
